+++ src/main/java/com/perennialsys/repository/LeaveRepository.java
@@ -1,12 +1,12 @@
 package com.perennialsys.repository;
 
 import com.perennialsys.entity.Leave;
-import com.perennialsys.entity.LeaveBalance;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LeaveRepository extends JpaRepository<Leave,Long> {
+public interface LeaveRepository extends JpaRepository<Leave, Integer> {
+
 
 
 }
